Removing unlink, adding attempt and adding docs
