+++ lib/patch_adams/core.rb
@@ -3,7 +3,6 @@ require_relative './minimal'
 require_relative './core/array/assign_numbers'
 require_relative './core/array/to_csv'
 require_relative './core/array/to_h'
-require_relative './core/file/unlink'
 require_relative './core/hash/fix_missing'
 require_relative './core/hash/diff'
 require_relative './core/hash/dig'
@@ -12,7 +11,6 @@ require_relative './core/hash/replace_key'
 require_relative './core/json/pretty_parse'
 require_relative './core/module/required'
 require_relative './core/module/revamp'
-require_relative './core/object/defaulting'
 require_relative './core/object/ensure_present'
 require_relative './core/object/to_stringified_h'
 require_relative './core/regexp/matches'
@@ -20,8 +18,10 @@ require_relative './core/yaml/load_interpolated'
 
 require_relative './core/as_module_name'
 require_relative './core/as_variable_name'
+require_relative './core/attempt'
 require_relative './core/captures'
 require_relative './core/configurable'
+require_relative './core/defaulting'
 require_relative './core/like'
 require_relative './core/status'
 require_relative './core/to_struct'
