+++ lib/patch_adams/core/defaulting.rb
@@ -1,0 +1,11 @@
+class NilClass
+  def defaulting(default_value)
+    default_value
+  end
+end
+
+class Object
+  def defaulting(_default_value)
+    self
+  end
+end
