--- lib/patch_adams/core/file/unlink.rb
@@ -1,4 +1,0 @@
-class File
-  def unlink
-  end
-end
