+++ spec/patch_adams/defaulting_spec.rb
@@ -1,0 +1,6 @@
+require_relative '../spec_helper'
+
+describe '#defaulting' do
+  it { expect(5.defaulting(4)).to eq 5 }
+  it { expect(nil.defaulting(4)).to eq 4 }
+end
